Added font and provider app state
