+++ doc/confapp_flutter/pubspec.yaml
@@ -79,12 +79,12 @@ flutter:
   # "family" key with the font family name, and a "fonts" key with a
   # list giving the asset and other descriptors for the font. For
   # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
+  fonts:
+    - family: RobotoMono
+      fonts:
+        - asset: fonts/Roboto_Mono/static/RobotoMono-Regular.ttf
+        - asset: fonts/Roboto_Mono/static/RobotoMono-Bold.ttf
+          weight: 700
   #   - family: Trajan Pro
   #     fonts:
   #       - asset: fonts/TrajanPro.ttf
